+++ fabric/src/main/java/io/github/sakurawald/fuji/core/auxiliary/minecraft/CommandHelper.java
@@ -124,6 +124,11 @@ public class CommandHelper {
     }
 
     public static void updateCommandTree() {
+        // NOTE: No need to update if the command manager is not initialized.
+        if (ServerHelper.getServer() == null) {
+            return;
+        }
+
         CommandManager commandManager = ServerHelper.getServer().getCommandManager();
         PlayerHelper.Lookup
             .getOnlinePlayers()